backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pytesseract
SpeechRecognition
pydub
